requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-jose
passlib
dotenv
bcrypt==4.1.2
python-multipart